pyproject.toml
[build-system]
requires      = ["setuptools>=75.1.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "onyxgenai"
version = "1.0.5"
description = "Python SDK for interacting with Onyx Generative AI Services"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "requests",
    "spacy",
    "PyPDF2",
    "python-pptx",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.urls]
Homepage = "https://github.com/MetroStar/onyx-genai-sdk"

[tool.setuptools]
packages.find.include = ["onyxgenai"]
packages.find.exclude = ["tests"]

[tool.ruff]
select = ["B", "C4", "EXE", "F", "E", "ISC", "ICN", "INP", "PIE", "SIM", "W", "T20", "UP", "T10", "G", "C90", "ERA"]
ignore = ["B008", "SIM102", "T201"]
fixable = ["F", "E", "B", "C4", "EXE", "ISC", "ICN", "INP", "PIE", "SIM", "W", "T20", "UP"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".mypy_cache",
    ".pre-commit-cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "venv",
    "docs",
    "__pycache",
    "**/migrations/*",
    "**/notebooks/*",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Assume Python 3.11.
target-version = "py311"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
